separated task and column
