+++ package.json
@@ -15,7 +15,7 @@
     "react-beautiful-dnd": "^12.2.0",
     "react-dom": "^16.12.0",
     "react-router-dom": "^5.1.2",
-    "react-scripts": "3.4.0",
+    "react-scripts": "^3.4.3",
     "typeface-roboto": "0.0.75",
     "uuid": "^7.0.0"
   },
